Fix Driver & Place Dao

+++ blablacar-persistence/src/main/java/cz/fi/muni/pa165/teamred/dao/PlaceDaoImpl.java
@@ -20,13 +20,19 @@ public class PlaceDaoImpl implements PlaceDao {
 
     @Override
     @Transactional
-    public void create(Place place) {
+    public void create(Place place) throws IllegalArgumentException {
+        if (place == null) {
+            throw new IllegalArgumentException("Place is null.");
+        }
         em.persist(place);
     }
 
     @Override
     @Transactional
-    public void delete(Place place) {
+    public void delete(Place place) throws IllegalArgumentException {
+        if (place == null) {
+            throw new IllegalArgumentException("Place is null.");
+        }
         em.remove(em.contains(place) ? place : em.merge(place));
     }
 
@@ -38,13 +44,20 @@ public class PlaceDaoImpl implements PlaceDao {
 
     @Override
     @Transactional
-    public Place findById(Long id) {
+    public Place findById(Long id) throws IllegalArgumentException {
+        if (id == null) {
+            throw new IllegalArgumentException("Id is null.");
+        }
         return em.find(Place.class, id);
     }
 
     @Override
     @Transactional
-    public Place findByName(String name) {
+    public Place findByName(String name) throws IllegalArgumentException {
+        if (name == null) {
+            throw new IllegalArgumentException("Name is null.");
+        }
+
         try {
             return em.createQuery("select p from Place p where p.name = :name", Place.class)
                     .setParameter("name", name)
